# API-Documentation.yaml
openapi: 3.0.3
info:
  title: VFS Bank Microservices API
  description: |
    VFS Bank (Veteran Financial Service Bank) full-stack digital banking platform API documentation.
    
    ## Architecture
    - **API Gateway**: http://localhost:8080 (all requests go through here)
    - **Auth Service**: http://localhost:8081
    - **Account Service**: http://localhost:8082
    - **Loan Service**: http://localhost:8083
    - **FD/RD Service**: http://localhost:8084
    - **Eureka Discovery**: http://localhost:8761
    
    ## Authentication
    Most endpoints require a JWT Bearer token in the `Authorization` header.
    1. Call `POST /api/auth/login` to get an access token
    2. Include header: `Authorization: Bearer <token>`
    
    ## Default Users (seeded by DataInitializer)
    | Username | Password | Role |
    |---|---|---|
    | john.doe | Secure@123 | CUSTOMER |
    | loan.officer | Secure@123 | LOAN_OFFICER |
    | branch.manager | Secure@123 | BRANCH_MANAGER |
    | admin.user | Secure@123 | SYSTEM_ADMIN |
  version: 1.0.0
  contact:
    name: Banking System Team
servers:
  - url: http://localhost:8080
    description: API Gateway (Production)
  - url: http://localhost:8081
    description: Auth Service (Direct)
  - url: http://localhost:8082
    description: Account Service (Direct)
  - url: http://localhost:8083
    description: Loan Service (Direct)
  - url: http://localhost:8084
    description: FD/RD Service (Direct)
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with credentials
      description: Authenticates user and returns JWT access + refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: john.doe
              password: Secure@123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Creates a new user account. Defaults to CUSTOMER role if not specified.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: new.user
              password: Secure@123
              email: newuser@test.com
              role: CUSTOMER
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Username or email already exists
  /api/auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a valid refresh token for a new access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          description: Invalid refresh token
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      description: Changes the authenticated user's password.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Current password is incorrect
  /api/auth/profile:
    put:
      tags: [Auth]
      summary: Update user profile
      description: Updates email or employee ID of the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              email: john.doe.updated@test.com
              employeeId: EMP1001
      responses:
        '200':
          description: User profile updated successfully
        '400':
          description: Validation error or email already in use
  /api/auth/users/{id}:
    put:
      tags: [Auth]
      summary: Update user details by ID
      description: Updates email, employee ID, status, or role of a specific user.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              email: updated.user@test.com
              status: ACTIVE
              role: CUSTOMER
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Invalid request or user not found

  # ==================== ACCOUNTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/auth/login
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: john.doe
        password:
          type: string
          example: Secure@123
        loginType:
          type: string
          enum: [CUSTOMER, EMPLOYEE, ADMIN]
    RegisterRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: new.user
        password:
          type: string
          minLength: 8
          example: Secure@123
        email:
          type: string
          format: email
          example: newuser@test.com
        role:
          type: string
          description: Optional, defaults to CUSTOMER
          enum: [CUSTOMER, LOAN_OFFICER, BRANCH_MANAGER, SYSTEM_ADMIN]
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PasswordChangeRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        employeeId:
          type: string
        status:
          type: string
          example: ACTIVE
        role:
          type: string
          example: CUSTOMER
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Login successful
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: integer
              example: 86400000
            requiresTwoFactor:
              type: boolean
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                email:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
                permissions:
                  type: array
                  items:
                    type: string
    TokenRefreshResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            tokenType:
              type: string
            expiresIn:
              type: integer

    # ---- Accounts ----
...